:heartbeat:添加Alibaba Druid连接池，支持数据池监控，SQL防注入；基于Quartz封装了定时任务调度暂用于ERROR日志告警 下面需要支持告警推送至开发者邮箱；更新Readme

+++ heathy-server-bean/src/main/java/com/zcy/cn/vo/AnnotationUser.java
@@ -15,5 +15,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class AnnotationUser {
     private String openId;
-    private Integer uId;
+    private Long uId;
 }
